requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
torchvision>=0.15.0
Pillow>=9.0.0
google-generativeai>=0.8.0
flask>=3.0.0
sounddevice>=0.4.6
paho-mqtt>=1.6.1,<2.0.0
